Fixes for lightmap + new single-pass terrain renderer
Additive effects color (from things like landing zone lights, explosion effects, etc) was effectively not visible or lighting terrain when using the new single-pass renderer. This was a byproduct of how most lighting calculations have moved to the shader, and the lightmap value (and how additive effect lighting was accumulated) was being calculated differently.

This adjusts the lightmap to now always be 4 channels. When using the new single-pass shaders, the lightmap contains:
1. Additive color from effects (etc) in the .rgb channels
2. The calculated ambient occlusion value in the 4th (.a) channel

The shaders can then handle these two values separately.

(When using the old fallback terrain shaders, the first 3 channels contain the old calculated fully-inclusive-but-not-the-most-accurate illumination value. Only minor changes were required to these shaders to ignore the 4th channel.)

+++ src/map.h
@@ -74,7 +74,7 @@ struct MAPTILE
 	// DISPLAY ONLY (NOT for use in game calculations)
 	uint8_t         ground;                 ///< The ground type used for the terrain renderer
 	uint8_t         illumination;           // How bright is this tile? = diffuseSunLight * ambientOcclusion
-	float			ambientOcclusion;		// ambient occlusion. from 1 (max occlusion) to 254 (no occlusion), similar to illumination.
+	uint8_t			ambientOcclusion;		// ambient occlusion. from 1 (max occlusion) to 254 (no occlusion), similar to illumination.
 	float           level;                  ///< The visibility level of the top left of the tile, for this client. for terrain lightmap
 	PIELIGHT        colour;					// color in terrain lightmap, based on tile.level and near light sources
 };
